Start implementation of a LaserScanPublisher and make use of it in the Hokuyo tutorial.

+++ android_hokuyo/src/org/ros/rosjava/android/hokuyo/Decoder.java
@@ -17,13 +17,14 @@ public class Decoder {
     return high + mid + low;
   }
 
-  public static List<Integer> decode(String buffer, int blockSize) {
+  public static List<Float> decode(String buffer, int blockSize) {
     Preconditions.checkArgument(blockSize == 3);
     Preconditions.checkArgument(buffer.length() % blockSize == 0);
-    List<Integer> data = Lists.newArrayList();
+    List<Float> data = Lists.newArrayList();
     for (int i = 0; i < buffer.length(); i += blockSize) {
       if (blockSize == 3) {
-        data.add(decode3Letter(buffer.substring(i, i + 3)));
+        // sensor_msgs/LaserScan uses floats for ranges.
+        data.add((float) decode3Letter(buffer.substring(i, i + 3)));
       }
     }
     return data;
